requirements.txt
requests==2.26.0
python-dateutil==2.8.1
Jinja2==3.0.3
